separating httpserver instantiation from routes
